Moved mocks to Mocks folder.
